feat(balances): make public only neccessary selectors, refactor app to use only those
