--- packages/blockchain-wallet-v4-frontend/src/components/Balances/total/selectors.ts
@@ -1,27 +1,0 @@
-import { lift } from 'ramda'
-
-import { fiatToString } from '@core/exchange/utils'
-import { createDeepEqualSelector } from '@core/utils'
-import { selectors } from 'data'
-
-import * as S from '../selectors'
-
-export const getTotalBalance = createDeepEqualSelector(
-  [
-    S.getFiatBalanceInfo,
-    S.getCoinsBalanceInfo,
-    selectors.core.settings.getCurrency,
-    selectors.router.getPathname
-  ],
-  (fiatBalanceInfoR, coinsBalanceInfo, currency, path) => {
-    const transform = (fiatBalance, coinsBalanceInfo, currency) => {
-      const balanceReducer = (acc, cur) => acc + Number(cur.getOrElse('0'))
-      const coinsBalance = coinsBalanceInfo.reduce(balanceReducer, 0)
-      const total = Number(fiatBalance) + coinsBalance
-      const totalBalance = `${fiatToString({ unit: currency, value: total })}`
-      return { path, totalBalance }
-    }
-
-    return lift(transform)(fiatBalanceInfoR, coinsBalanceInfo, currency)
-  }
-)
