--- packages/blockchain-wallet-v4-frontend/src/data/coins/selectors/balances/index.ts
@@ -1,4 +1,0 @@
-export * from './custodial.selectors'
-export * from './fiat.selectors'
-export * from './non-custodial.selectors'
-export * from './total.selectors'
