+++ packages/blockchain-wallet-v4-frontend/src/data/coins/selectors/coins/btc.tsx
@@ -7,7 +7,7 @@ import { ADDRESS_TYPES } from '@core/redux/payment/btc/utils'
 import { ExtractSuccess } from '@core/remote/types'
 import { CoinType } from '@core/types'
 import { createDeepEqualSelector } from '@core/utils'
-import { getCoinInterestBalance, getCoinTradingBalance } from 'data/coins/selectors'
+import { selectors } from 'data'
 import { CoinAccountSelectorType } from 'data/coins/types'
 import { generateInterestAccount, generateTradingAccount } from 'data/coins/utils'
 import { SwapAccountType } from 'data/types'
@@ -27,8 +27,8 @@ export const getAccounts = createDeepEqualSelector(
     coreSelectors.wallet.getHDAccounts, // non-custodial accounts
     coreSelectors.data.btc.getAddresses, // non-custodial xpub info
     coreSelectors.common.btc.getActiveAddresses, // imported addresses
-    (state, { coin }) => getCoinTradingBalance(coin, state), // custodial accounts
-    (state, { coin }) => getCoinInterestBalance(coin, state), // custodial accounts
+    (state, { coin }) => selectors.balances.getCoinTradingBalance(coin, state), // custodial accounts
+    (state, { coin }) => selectors.balances.getCoinInterestBalance(coin, state), // custodial accounts
     (state, ownProps): CoinAccountSelectorType & { coin: CoinType } => ownProps // selector config
   ],
   (btcAccounts, btcDataR, importedAddressesR, sbBalanceR, interestBalanceR, ownProps) => {
