+++ packages/blockchain-wallet-v4-frontend/src/data/coins/selectors/coins/erc20.tsx
@@ -6,7 +6,7 @@ import { coreSelectors } from '@core'
 import { BSBalanceType } from '@core/network/api/buySell/types'
 import { ExtractSuccess } from '@core/remote/types'
 import { createDeepEqualSelector } from '@core/utils'
-import { getCoinTradingBalance } from 'data/coins/selectors'
+import { selectors } from 'data'
 import { generateTradingAccount } from 'data/coins/utils'
 import { SwapAccountType, SwapBaseCounterTypes } from 'data/types'
 
@@ -60,7 +60,7 @@ export const getAccounts = createDeepEqualSelector(
   [
     coreSelectors.kvStore.eth.getDefaultAddress,
     (state, { coin }) => coreSelectors.data.eth.getErc20Balance(state, coin), // non-custodial metadata
-    (state, { coin }) => getCoinTradingBalance(coin, state), // custodial accounts
+    (state, { coin }) => selectors.balances.getCoinTradingBalance(coin, state), // custodial accounts
     (state, ownProps) => ownProps // selector config
   ],
   (ethAddressR, erc20BalanceR, sbBalanceR, ownProps) => {
