+++ packages/blockchain-wallet-v4-frontend/src/data/coins/selectors/coins/eth.tsx
@@ -7,7 +7,7 @@ import { BSBalanceType } from '@core/network/api/buySell/types'
 import { ADDRESS_TYPES } from '@core/redux/payment/btc/utils'
 import { ExtractSuccess } from '@core/remote/types'
 import { createDeepEqualSelector } from '@core/utils'
-import { getCoinTradingBalance } from 'data/coins/selectors'
+import { selectors } from 'data'
 import { generateTradingAccount } from 'data/coins/utils'
 import { SwapAccountType } from 'data/types'
 
@@ -26,7 +26,7 @@ export const getAccounts = createDeepEqualSelector(
   [
     coreSelectors.data.eth.getAddresses, // non-custodial accounts
     coreSelectors.kvStore.eth.getAccounts, // non-custodial metadata
-    (state, { coin }) => getCoinTradingBalance(coin, state), // custodial accounts
+    (state, { coin }) => selectors.balances.getCoinTradingBalance(coin, state), // custodial accounts
     (state, ownProps) => ownProps // selector config
   ],
   (ethDataR, ethMetadataR, sbBalanceR, ownProps) => {
