+++ packages/blockchain-wallet-v4-frontend/src/data/sagas.js
@@ -9,7 +9,6 @@ import * as preferences from './preferences/sagas'
 import * as prices from './prices/sagas'
 import * as router from './router/sagas'
 import * as session from './session/sagas'
-import * as signup from './signup/sagas'
 import * as wallet from './wallet/sagas'
 
 export {
