+++ packages/blockchain-wallet-v4-frontend/src/data/selectors.ts
@@ -2,6 +2,7 @@ import { coreSelectors as core } from '@core'
 
 import * as alerts from './alerts/selectors'
 import * as auth from './auth/selectors'
+import * as balances from './balances/selectors'
 import * as cache from './cache/selectors'
 import * as coins from './coins/selectors'
 import * as components from './components/selectors'
@@ -20,6 +21,7 @@ import * as signup from './signup/selectors'
 export {
   alerts,
   auth,
+  balances,
   cache,
   coins,
   components,
