+++ packages/blockchain-wallet-v4-frontend/src/layouts/Wallet/MenuLeft/Balances/WalletBalance/index.tsx
@@ -3,7 +3,6 @@ import { connect, ConnectedProps } from 'react-redux'
 import { prop } from 'ramda'
 import { bindActionCreators } from 'redux'
 
-import { getCoinsSortedByBalance } from 'components/Balances/selectors'
 import { actions, selectors } from 'data'
 
 import Template from './template'
@@ -28,7 +27,7 @@ class WalletBalanceContainer extends React.PureComponent<Props> {
 }
 
 const mapStateToProps = (state) => ({
-  coins: getCoinsSortedByBalance(state).getOrElse([]),
+  coins: selectors.balances.getTotalWalletBalancesSorted(state).getOrElse([]),
   totalBalancesDropdown: selectors.preferences.getTotalBalancesDropdown(state)
 })
