+++ packages/blockchain-wallet-v4-frontend/src/scenes/Home/Holdings/CoinBalance/selectors.ts
@@ -1,9 +1,9 @@
-import * as balanceSelectors from 'components/Balances/selectors'
+import { selectors } from 'data'
 
 import { OwnProps } from '.'
 
 export const getData = (state, ownProps: OwnProps) => {
-  return balanceSelectors.getBalanceSelector(ownProps.coin)(state)
+  return selectors.balances.getCoinTotalBalance(ownProps.coin)(state)
 }
 
 export default getData
